Use node's util.types.isUint8Array etc for isTypedArray
isTypedArray is implemented with `instanceof`, which does not work in jest (https://github.com/facebook/jest/issues/11864). Instead, use node's builtin `util.types.isUint8Array`, `util.types.isFloat32Array`, etc to perform this check.

Fixes #7175.
This may also address #7064, but it does not fix the root cause.
