+++ tfjs-core/src/platforms/platform_browser.ts
@@ -90,6 +90,12 @@ export class PlatformBrowser implements Platform {
       }, true);
     }
   }
+
+  isTypedArray(a: unknown): a is Uint8Array | Float32Array | Int32Array
+    | Uint8ClampedArray {
+    return a instanceof Float32Array || a instanceof Int32Array ||
+      a instanceof Uint8Array || a instanceof Uint8ClampedArray;
+  }
 }
 
 if (env().get('IS_BROWSER')) {
